LayoutTests/platform/gtk/imported/w3c/web-platform-tests/html/cross-origin-opener-policy/coop-csp-sandbox.https-expected.txt

CONSOLE MESSAGE: Navigation was blocked by Cross-Origin-Opener-Policy
CONSOLE MESSAGE: Navigation was blocked by Cross-Origin-Opener-Policy

PASS CSP: sandbox allow-popups allow-scripts allow-same-origin; CSP sandboxed Cross-Origin-Opener-Policy popup should result in a network error
PASS CSP: sandbox allow-popups allow-scripts; CSP sandboxed Cross-Origin-Opener-Policy popup should result in a network error
